Renamed: spec_helper to rails_helper refs #26
