+++ spec/controllers/service_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe ServiceController, :type => :controller  do
   include Devise::TestHelpers
