+++ spec/controllers/welcome_controller_spec.rb
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-require File.dirname(__FILE__) + '/../spec_helper'
+require File.dirname(__FILE__) + '/../rails_helper'
 
 describe WelcomeController, :type => :controller do
   include Devise::TestHelpers
