+++ spec/models/complex_graph_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe ComplexGraph do
   pending "add some examples to (or delete) #{__FILE__}"
